backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.2
colorama==0.4.6
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.2.6
opencv-python==4.12.0.88
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
